refactor(reactotron-mcp): extract shared applyRedaction/getClientRedactionConfig into redaction.ts
Removes duplicate definitions from tools.ts and resources.ts.
