+++ lib/reactotron-mcp/src/redaction.ts
@@ -1,4 +1,5 @@
 import type { McpRedactionRules, McpRedactionConfig } from "reactotron-core-contract"
+import type ReactotronServer from "reactotron-core-server"
 
 export const REDACTED = "[REDACTED]"
 
@@ -87,6 +88,36 @@ export function resolveEffectiveRules(
   }
 
   return rules
+}
+
+/**
+ * Get the client's McpRedactionConfig from the connection object.
+ * Uses the first connected app's config (single-app case) or undefined if none.
+ */
+export function getClientRedactionConfig(server: ReactotronServer, clientId?: string): McpRedactionConfig | undefined {
+  const connections = server.connections as any[]
+  if (clientId) {
+    const conn = connections.find((c) => c.clientId === clientId)
+    return conn?.mcpRedaction
+  }
+  if (connections.length === 1) {
+    return connections[0]?.mcpRedaction
+  }
+  return undefined
+}
+
+/** Apply redaction to data based on server config and connected client config. */
+export function applyRedaction(
+  data: unknown,
+  server: ReactotronServer,
+  serverRedactionConfig: McpRedactionServerConfig,
+  clientId?: string,
+  basePath = ""
+): unknown {
+  const clientConfig = getClientRedactionConfig(server, clientId)
+  const rules = resolveEffectiveRules(serverRedactionConfig, clientConfig)
+  if (!rules) return data
+  return redact(data, rules, basePath)
 }
 
 function deepCopyRules(rules: McpRedactionRules): McpRedactionRules {
